Move the f5 related stuff out to a separate class
Improves readability somewhat. After I work with this a bit I think
it'll end up getting moved into the f5-icontrol gem instead.

There's an ugly hack where I stub out the Chef library loading. What
would happen is that the mocks would be created, but the convergence
would reload the class and lose the mocks. So I manually require the
library, mock it, then run Chef.

When the behaviour is extracted to the f5-icontrol gem this
embarrassment will go away.