Install packages at compile time instead of execute
Chef runs in two passes -- first it compiles the recipe then it executes
it. `chef_gem` runs in the first, `package` runs in the second. Thus, if
something like `libxml2-devel` isn't installed, it won't get installed
before the gem.

This change pushes the packages to the compile step by using some
behaviour inside build-essential (which we need anyway). See
https://github.com/opscode-cookbooks/build-essential#gems-with-c-extensions
for more information.