Force the right version of the gem; TK fixes
Was running into problems where older nodes could not use the new
behaviour in the f5-icontrol gem because it wasn't getting upgraded.

Also working on test kitchen tests. Note that there are some items that
don't go with the project, such as the credentials to my lab load
balancer ;)