A pile of modernization (#5)
* Remove the fil in the blank license section

Signed-off-by: Tim Smith <tsmith@chef.io>

* Rework some of the wording in the readme

Signed-off-by: Tim Smith <tsmith@chef.io>

* Speed things up with multi package installs

Signed-off-by: Tim Smith <tsmith@chef.io>

* Always use inline resources

Signed-off-by: Tim Smith <tsmith@chef.io>

* Remove extra state updating and duplicate logging

Converge_by handles this all

Signed-off-by: Tim Smith <tsmith@chef.io>

* Remove the default recipe

Go 100% resource

Signed-off-by: Tim Smith <tsmith@chef.io>

* Require Chef 12.1+

Require a mildly modern version of Chef

Signed-off-by: Tim Smith <tsmith@chef.io>

* Remove the test cookbook readme

It doesn't really add anything

Signed-off-by: Tim Smith <tsmith@chef.io>

* Document the attribute and don't use the word LWRP

We're trying to kill off the LWRP acronym as it doesn't mean much to most people.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Resolve foodcritic and cookstyle warnings

Signed-off-by: Tim Smith <tsmith@chef.io>

* Avoiding using the Chef namespace for a class

This is going to end poorly if we implement a class with this name in Chef

Signed-off-by: Tim Smith <tsmith@chef.io>

* Allow specifying the data bag name as an attribute

It's pretty common to want to use a different name for a data bag. This makes that possible and fixes the incorrect variable names since databag_name was the variable for data bag items and not the actual bags.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Remove readme from the test cookbook metadata

Signed-off-by: Tim Smith <tsmith@chef.io>

* Part 1 in removing the need for build-essential

Install the deps that nokogiri needs. New nokogiri doesn't actually require libxml anymore either so we can skip this. We also need to make sure that we compile time install the gem

Signed-off-by: Tim Smith <tsmith@chef.io>

* Remove some extra spaces

Signed-off-by: Tim Smith <tsmith@chef.io>

* Use the new supermarket URL

Signed-off-by: Tim Smith <tsmith@chef.io>

* Improve credential passing

Add additional logging in the credential access
Add an empty attribute for credentials to the attributes file
Allow passing creds directly on the resource
Raise a helpful error if none of the cred access methods were successful
Use environmental variable creds set to node attributes in test kitchen for simpler testing

Signed-off-by: Tim Smith <tsmith@chef.io>

* Remove build-essential dep and only support RHEL for now

Debian support is coming next

Signed-off-by: Tim Smith <tsmith@chef.io>

* Make sure to install patch as well

Gem installs fail w/o this

Signed-off-by: Tim Smith <tsmith@chef.io>

* Use the latest gem release

Signed-off-by: Tim Smith <tsmith@chef.io>

* Create the pool before the VIP so it doesn't fail

Signed-off-by: Tim Smith <tsmith@chef.io>

* Convert to custom resources

Signed-off-by: Tim Smith <tsmith@chef.io>

* Require Chef 12.7+

This fixes some issues with custom resources

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add support for Debian/Suse and extract the gem install away

This avoids seeing package/gem resources in every use of the resource. The first usage will fail to load the gem and then fallback to package/gem install. From then on it just loads the gem and continues on to create the client

Signed-off-by: Tim Smith <tsmith@chef.io>