Support using an FQDN for the VIP (experimental) (#15)
It would be nice to be able to get IP management out of Chef and into
something that's good at it, such as DNS.

Now you can create a vip such as

```Ruby
f5_vip 'foo' do
  address "origin-#{node.chef_environment}.example.com"
  ...
end
```

And it'll create the VIP with the resolved address, but only if the
resolution succeeds.

I'm positive there are some untouched corner cases, such as if you have
multiple A records. Or if you change the address, this may not react
well. However for the base case, this seems to work well. We've been
using this method for several months now, and I'm moving the code to
the resource itself.